+++ src/components/index.js
@@ -7,12 +7,14 @@ export { default as InfoBox } from "./info-box/index";
 export { default as FieldGroup } from "./field-group/index";
 export { default as Icon } from "./icons/index";
 export { default as InlineSelect } from "./inline-select/index";
+export { default as Select } from "./select/index";
 export { default as Link } from "./link/index";
 export { default as Loader } from "./loader/loader";
 export { default as Popover } from "./popover/index";
 export { default as Textarea } from "./textarea/index";
-export { default as ThemeProviderWrapper } from "./theme/index";
+export { ThemeProviderWrapper, colors } from "./theme/index";
 export { Typography, Header, Paragraph } from "./typography/index";
+export { Toggle, ToggleOption } from "./toggle/index";
 export {
   Input,
   InputStatus,
@@ -22,4 +24,3 @@ export {
 } from "./input/index";
 export { Table, Td, Tr } from "./table/index";
 export { Tab, Tabs } from "./tabs/index";
-export { colors } from "./theme/index";