+++ src/components/theme/index.js
@@ -1,2 +1,2 @@
-export { default } from "./theme-provider";
+export { default as ThemeProviderWrapper } from "./theme-provider";
 export { default as colors } from "./base/colors";