+++ src/components/toggle/index.js
@@ -1,0 +1,2 @@
+export { default as Toggle } from "./toggle";
+export { default as ToggleOption } from "./toggle-option";